requirements.txt
discord.py
asyncpg
Flask
gunicorn